이동 공격 이상한거 수정ㄹ

+++ lib/Type/Type.h
@@ -30,9 +30,9 @@ CPacket& operator>>(CPacket& packet, FVector& vec);
 
 struct FRotator
 {
-	float Pitch;
-	float Yaw;
-	float Roll;
+	double Pitch;
+	double Yaw;
+	double Roll;
 };
 
 CPacket& operator<<(CPacket& packet, FRotator& rot);
